+++ scripts/slugify.js
@@ -1,0 +1,3 @@
+const slugify = require("slugify");
+
+module.exports = str => slugify(str, { remove: /['':;,]/g, lower: true });
