+++ src/partials/dribbble.ejs
@@ -1,0 +1,28 @@
+<% const {
+  description,
+  published_at,
+  html_url,
+  images: { normal },
+  title
+} = item %>
+
+<li class="item">
+  <a class="item__img" href="<%- html_url %>">
+    <img
+      src="<%- normal %>"
+      alt="<%- title %>"
+      title="<%- title %> on Dribbble"
+      width="400"
+      height="300"
+    />
+  </a>
+  <div class="item__content">
+    <h3 class="title">
+      <a href="<%- html_url %>" class="link">
+        <%- title %>
+      </a>
+    </h3>
+    <p class="byline"><%- published_at %></p>
+    <div class="description"><%- description %></div>
+  </div>
+</li>
