+++ src/partials/employment.ejs
@@ -1,0 +1,51 @@
+<% const {
+  byline,
+  company,
+  companyUrl,
+  description,
+  projects,
+  title,
+  thumbnail // auto-generated
+} = item %>
+
+<li class="item">
+  <div class="item__img">
+    <img src="/assets/img/employment/<%- thumbnail %>" width="70" height="70" />
+  </div>
+  <div class="item__content">
+    <h3 class="title">
+      <%- title %> @ <a href="<%- companyUrl %>" class="link"><%- company %></a>
+    </h3>
+    <p class="byline"><%- byline %></p>
+    <p class="description"><%- description %></p>
+
+    <% if (projects) { %>
+    <ul class="projects clearfix">
+      <% projects.forEach(project => { %>
+      <li>
+          <h4>
+            Project: <%- project.title %>
+          </h4>
+
+          <p class="description">
+            <%- project.description %>
+            Tools: <%- project.tools %>.
+          </p>
+
+          <% if (project.links) { %>
+            <ul class="links">
+              <% project.links.forEach(link => { %>
+              <li>
+                <a href="<%- link.url %>" class="link">
+                  <%- link.title %>
+                </a>
+              </li>
+              <% }) %>
+            </ul>
+          <% } %>
+      </li>
+      <% }) %>
+    </ul>
+    <% } %>
+  </div>
+</li>
