+++ src/partials/publishings.ejs
@@ -1,0 +1,27 @@
+<% const {
+  date,
+  link,
+  publisher,
+  readerFav,
+  thumbnail,
+  title,
+  tweets
+} = item %>
+
+<li class="item">
+  <a class="item__img" href="<%- link %>">
+    <img src="assets/img/publishings/<%- thumbnail %>" width="70" height="70" />
+  </a>
+  <div class="item__content">
+    <h3 class="title link"><a href="<%- link %>"><%- title %></a></h3>
+    <p class="byline">
+      <time><%- date %></time> via <%- publisher %>
+      <% if (readerFav) { %>
+        <em class="readers-fav"> readers’ fav</em>
+      <% } %>
+    </p>
+    <% if (tweets) { %>
+      <%- include("./tweets.ejs", { tweets }) %>
+    <% } %>
+  </div>
+</li>
